log: have the compiler read the diagnostics against their formats
The two functions behind XPOST_LOG_ERR, WARN, INFO, DBG and DUMP take a
format string and are called from every part of the tree, and neither
carried the mark that has the compiler check the arguments against it.
So the interpreter's diagnostics were the one body of formatted calls
nobody read. Marking them finds five defects, and the compiler validates
the index pair against the parameter types, so a mistyped mark is
refused at the declaration rather than checking the wrong argument.

Three of them pass a size_t through %d, in the memory file's own growth
and truncation reports, at both object widths. One passes the object's
integer through %u in the array bounds report, at the wide width. Each
is now widened to the largest of its kind rather than cast down to fit
the conversion, since narrowing would answer the warning by discarding
the value it was warning about.

The fifth is the interesting one. A run that cannot find its boot files
says which directory it looked in -- except that the message carried no
conversion at all and the path sat behind it as an argument nobody read.
Six expansions of that macro, six diagnostics that named no directory.
It names one now.

And the format macros the wide object is printed through were wrong for
the width they exist to serve: the integer types are long long and
unsigned long long, and the conversions said long. Where a long is 64
bits the two coincide and nothing shows; where it is 32 -- which this
tree is built for -- a wide object's value would be read back through
half of itself. The macro is fixed rather than its call sites, which is
what having a macro is for.

Zero warnings at both widths in debug and in release.

Commit message generated by Claude (Anthropic).

Source code generated by Claude (Anthropic).

+++ src/lib/xpost_array.c
@@ -152,7 +152,8 @@ int xpost_array_put_memory(Xpost_Memory_File *mem,
         return ret;
     if (i >= a.comp_.sz)
     {
-        XPOST_LOG_ERR("cannot put value in array (rangecheck) %u >= [%u]", i, a.comp_.sz);
+        XPOST_LOG_ERR("cannot put value in array (rangecheck) %lld >= [%u]",
+                      (long long)i, a.comp_.sz);
         /*breakhere((Xpost_Context *)mem);*/
         return rangecheck;
     }
